Major refactor and improved lang support
This commit refactors the code to retrieve the currency by its ID
instead of searching for it within the known currencies by icon.

It also automatically gets the currency name, description, and quality
from the currency info object, not requiring to specify the name and
description in the translations anymore. This way, the addOn now
partialy supports all languages that WoW supports (just a few
translations will still need to be added manually).